Edited KinectStreams for raw x and y data.
Attempted editing of the kinectstreams code to obtain x and y data.  Should be slightly more accurate as the z coordinate (involving most of the estimation) has been omitted.  Note: This code is only to be used if we decide to change our record method to walking horizontally across the kinect's field of view.