Another Calculation Addition
staticDistances.py has functions for graphing and calculating the 'static' distances between joint points. There is a problem when graphing the animation function for all of the data. Does not run properly if the AllData animation graph is graphed before anything else. Everything works fine if the animation is graphed last, seemingly. Initial observations of the staticDistances seems like the data definitely needs to be smoothed somehow