More Python code and Adjusted KinectStreams
Modified KinectStreams to have a Button to take a single Frame Snapshot of a subject that is standing still directly in front of the Kinect. Also added a subjectID string variable in KinectStreams that needs to be modified before each subject to properly label their data files. Adjusted folder layout to have separate folders for FrontFacing data and SideFacing data. Have python code to graph single frame of frontfacing and have a method for calculating joint distance. I have not implemented an angle calculation for the front facing yet. There is a sample Front Facing CSV provided.