More Additions
KinectStreams now has functionality for side view recording, forward view recording, and forward view snapshots. Added if statements to prevent potential problems by only allowing one method of recording to work at once. Created subfolders in FrontViews for the Dynamic Data and the Static Data. The MaxTesting10-20.csv file was purely for testing very basic functionality. It does not contain any data that should be used to test the accuracy of calculation methods. frontGraphingDynamic.py has a 3D animation graphing functionality for forward facing recorded data. It also has a basic graph function that takes in an array of angle data or joint length data. There is also a jointLength function that returns an array of the given joint length over time. There is also an rms method, 'potentially', that is also included, but I'm not sure if I created it properly and should be looked at maybe. I did not get to making angle calculation functions. Everything seems to work for now. If there are any questions, feel free to ask. All functions are designed to be Modular and able to work with various joint points, etc.