Added Subject ID input to KinectStreams and Minor UI Tweaks
Added a textbox that allows the user to input a name or identifier for the subject being recorded. Simply enter the subject name or identifier and click the Save Subject ID button. A minor function was added to autoselect text in the textbox when the box is clicked into for ease of use. Since there was multiple buttons of similar use in KinectStreams, I put colored borders around everything to distinguish between the buttons functionalities and which button is used with which. When a subject ID is entered, it will be displayed in the Title of the KinectStreams Window. I also made it so that no recording can be done until an initial subjectID is entered. As always, if something doesn't seem to work properly or if there are any questions, let me know.